Wed, Nov 12, 2025, 4:18 PM -08:00

+++ interpreter/interpreter.py
@@ -1,8 +1,8 @@
 user_input = input("Expression: ")
 user_array = user_input.split(" ")
 # print(user_array)
-x = int(user_array[0])
-z = int(user_array[2])
+x = float(user_array[0])
+z = float(user_array[2])
 y = str(user_array[1])
 
 if y == "+": result = round(x + z, 1)
